backend/requirements.txt
astroid==2.2.5
autopep8==1.4.4
boto3==1.13.19
botocore==1.16.19
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.3
configparser==5.0.0
defusedxml==0.6.0
Django==2.2.10
django-allauth==0.39.1
django-cors-headers==3.0.2
django-rest-auth==0.9.5
djangorestframework==3.9.4
docutils==0.15.2
gunicorn==19.9.0
idna==2.8
isort==4.3.21
jmespath==0.10.0
lazy-object-proxy==1.4.1
mccabe==0.6.1
oauthlib==3.0.1
pep8==1.7.1
psycopg2==2.8.5
pycodestyle==2.5.0
Pygments==2.6.1
pylint==2.3.1
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
s3transfer==0.3.3
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.3
wrapt==1.11.2
